Added tax system for multiplayer and server member enabled games
resourcecheck: added a tax system for multiplayer and server member enabled games.
Essentially takes a hard coded 5% of earnings and distributed it evenly to the players.
This helps them keep enough money in their personal coffers to have enough
cash to buy an AI or two and transportation.

INFO: This whole commit is using the tax system parts of following bigger
commits of github user uamadman
https://github.com/uamadman/A3-Antistasi/commit/90612e1f34e5260ec86ef7c26dec150a54e41837
https://github.com/uamadman/A3-Antistasi/commit/0bf0084191dd0597986e4dd4e7e934d0f98054b9